Code cleanup
- Remove IRegistry
- Improve the description of RegistryReaderRuntimesVersions
- Remove incorrectly assigned null types
- Rename property names to better describe the purpose.
- Rename RegistryReaderNetCoreRuntimesVersions to NetCoreRuntimeVersionsRegistryReader
